Move trino.formatter.ts to languages/trino/ dir
